Split the time code into new file and made a more RESTful endpoint
